fix: improve browser history navigation (back, forward)
